Update README.md
Business objective: Develop a a pricing model that can effectively predict the price of used cars and can help the business in devising profitable strategies using differential pricing.
Business in question: Used-Car, LLC
Business Location: India

Objectives:
- Explore and visualize the dataset
- Build a linear regression model to predict the prices of used cars
- Generate a set of insights and recommendations that will help the business

Data Directory:

S.No.: Serial Number

Name: Name of the car which includes Brand name and Model name

Location: The location in which the car is being sold or is available for purchase Cities

Year: Manufacturing year of the car

Kilometers_driven: The total kilometers driven in the car by the previous owner(s) in KM.

Fuel_Type: The type of fuel used by the car. (Petrol, Diesel, Electric, CNG, LPG)

Transmission: The type of transmission used by the car. (Automatic / Manual)

Owner: Type of ownership

Mileage: The standard mileage offered by the car company in kmpl or km/kg

Engine: The displacement volume of the engine in CC.

Power: The maximum power of the engine in bhp.

Seats: The number of seats in the car.

New_Price: The price of a new car of the same model in INR Lakhs.(1 Lakh = 100, 000)

Price: The price of the used car in INR Lakhs (1 Lakh = 100, 000)